# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bot Status Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type BotStatus = typeof BotStatus[keyof typeof BotStatus];


export const BotStatus = {
  online: 'online',
  offline: 'offline',
  error: 'error',
  starting: 'starting',
} as const;

export interface Bot {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  status: BotStatus;
  uptimeSeconds: number;
  messagesProcessed: number;
  commandsHandled: number;
  errorsToday: number;
  /** @nullable */
  lastSeenAt: string | null;
  createdAt: string;
}

export interface BotApiKeyInfo {
  botId: number;
  apiKey: string;
  createdAt: string;
}

export interface BotInput {
  /** @minLength 1 */
  name: string;
  description?: string;
}

export type BotUpdateStatus = typeof BotUpdateStatus[keyof typeof BotUpdateStatus];


export const BotUpdateStatus = {
  online: 'online',
  offline: 'offline',
  error: 'error',
  starting: 'starting',
} as const;

export interface BotUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  status?: BotUpdateStatus;
}

export type PingInputStatus = typeof PingInputStatus[keyof typeof PingInputStatus];


export const PingInputStatus = {
  online: 'online',
  offline: 'offline',
  error: 'error',
  starting: 'starting',
} as const;

export interface PingInput {
  status: PingInputStatus;
  messagesProcessed?: number;
  commandsHandled?: number;
  errorsToday?: number;
}

export type BotEventType = typeof BotEventType[keyof typeof BotEventType];


export const BotEventType = {
  info: 'info',
  warning: 'warning',
  error: 'error',
  online: 'online',
  offline: 'offline',
} as const;

export interface BotEvent {
  id: number;
  botId: number;
  type: BotEventType;
  message: string;
  createdAt: string;
}

export interface DashboardSummary {
  totalBots: number;
  onlineBots: number;
  offlineBots: number;
  errorBots: number;
  totalMessages: number;
  totalCommands: number;
  totalErrors: number;
}

export interface ErrorResponse {
  error: string;
}