# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Bot Status Dashboard API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: bots
    description: Bot management
  - name: dashboard
    description: Dashboard summaries
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /bots:
    get:
      operationId: listBots
      tags: [bots]
      summary: List all bots
      description: Returns all bots with their current status and metrics
      responses:
        "200":
          description: List of bots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bot"
    post:
      operationId: createBot
      tags: [bots]
      summary: Register a bot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BotInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bot"

  /bots/{id}:
    get:
      operationId: getBot
      tags: [bots]
      summary: Get bot details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Bot details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bot"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateBot
      tags: [bots]
      summary: Update bot info or status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BotUpdate"
      responses:
        "200":
          description: Updated bot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bot"
    delete:
      operationId: deleteBot
      tags: [bots]
      summary: Remove a bot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /bots/{id}/ping:
    post:
      operationId: pingBot
      tags: [bots]
      summary: Record a bot heartbeat / ping
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PingInput"
      responses:
        "200":
          description: Updated bot after ping
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bot"
        "401":
          description: Unauthorized - invalid or missing API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /bots/{id}/key:
    get:
      operationId: getBotApiKey
      tags: [bots]
      summary: Get the API key for a bot (dashboard use only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: API key info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotApiKeyInfo"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /bots/{id}/events:
    get:
      operationId: listBotEvents
      tags: [bots]
      summary: Get recent events for a bot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recent events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BotEvent"

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Dashboard-level aggregates across all bots
      responses:
        "200":
          description: Summary stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Bot:
      type: object
      required: [id, name, status, uptimeSeconds, messagesProcessed, commandsHandled, errorsToday, lastSeenAt, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [online, offline, error, starting]
        uptimeSeconds:
          type: integer
        messagesProcessed:
          type: integer
        commandsHandled:
          type: integer
        errorsToday:
          type: integer
        lastSeenAt:
          type: ["string", "null"]
        createdAt:
          type: string

    BotApiKeyInfo:
      type: object
      required: [botId, apiKey, createdAt]
      properties:
        botId:
          type: integer
        apiKey:
          type: string
        createdAt:
          type: string

    BotInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string

    BotUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        status:
          type: string
          enum: [online, offline, error, starting]

    PingInput:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [online, offline, error, starting]
        messagesProcessed:
          type: integer
        commandsHandled:
          type: integer
        errorsToday:
          type: integer

    BotEvent:
      type: object
      required: [id, botId, type, message, createdAt]
      properties:
        id:
          type: integer
        botId:
          type: integer
        type:
          type: string
          enum: [info, warning, error, online, offline]
        message:
          type: string
        createdAt:
          type: string

    DashboardSummary:
      type: object
      required: [totalBots, onlineBots, offlineBots, errorBots, totalMessages, totalCommands, totalErrors]
      properties:
        totalBots:
          type: integer
        onlineBots:
          type: integer
        offlineBots:
          type: integer
        errorBots:
          type: integer
        totalMessages:
          type: integer
        totalCommands:
          type: integer
        totalErrors:
          type: integer

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string