# lib/api-zod/src/generated/types/botApiKeyInfo.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bot Status Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface BotApiKeyInfo {
  botId: number;
  apiKey: string;
  createdAt: string;
}